requirements.txt
flask==3.0.0
gunicorn==21.2.0
pillow==10.2.0
numpy==1.26.4
requests==2.31.0
